feat: adicionando middleware de validação na atualização de assistido
