+++ src/routes/assisted.routes.js
@@ -4,6 +4,7 @@ const { Router } = require('express');
 
 const CreateAssistedValidation = require('@middlewares/validations/CreateAssistedValidation');
 const ShowAssistedValidation = require('@middlewares/validations/ShowAssistedValidation');
+const UpdateAssistedValidation = require('@middlewares/validations/UpdateAssistedValidation');
 
 const assistedRouter = Router();
 
@@ -11,7 +12,11 @@ assistedRouter.post('/', CreateAssistedValidation, AssistedController.store);
 assistedRouter.get('/', AssistedController.index);
 // use ?type=[TYPE] after id as query param to set the type of search
 assistedRouter.get('/:id', ShowAssistedValidation, AssistedController.show);
-assistedRouter.put('/update/:id', AssistedController.update);
+assistedRouter.put(
+  '/update/:id',
+  UpdateAssistedValidation,
+  AssistedController.update,
+);
 assistedRouter.delete('/:id', AssistedController.destroy);
 
 // Rota para associar um membro a um assistido
